+++ etos/test/etos_test.dart
@@ -1,3 +1,5 @@
+import 'dart:async';
+
 import 'package:etos/etos.dart';
 import 'package:test/expect.dart';
 import 'package:test/scaffolding.dart';
@@ -7,16 +9,9 @@ Future<void> main() async {
 
   setUp(() {
     etos = Etos<int>(state: 0)
-      ..on<IncrementEvent>((event, getter, setter) async {
-        Future.delayed(Duration(seconds: 1));
-        setter(getter() + 1);
-      })
+      ..on<IncrementEvent>(IncrementEventHandler())
       ..on<AddEvent>(AddHandler())
-      ..on<DecrementEvent>(
-        (event, stateGetter, stateSetter) {
-          stateSetter(stateGetter() - 1);
-        },
-      );
+      ..on<DecrementEvent>(DecrementEventHandler());
   });
 
   test('simple test', () {
@@ -46,17 +41,14 @@ Future<void> main() async {
 
     etos.dispatch(DecrementEvent());
   });
-}
 
-class AddHandler {
-  Future<void> call(
-      AddEvent event, StateGetter<int> get, StateSetter<int> set) async {
-    for (var i = 0; i < event.number; i++) {
-      await Future.delayed(Duration(milliseconds: 100));
-      final newState = get() + 1;
-      set(newState);
-    }
-  }
+  test('canCall test', () async {
+    expect(etos.states, emitsInOrder([0, 1, 0]));
+
+    etos.dispatch(IncrementEvent());
+    etos.dispatch(DecrementEvent());
+    etos.dispatch(DecrementEvent());
+  });
 }
 
 class IncrementEvent {}
@@ -67,4 +59,33 @@ class AddEvent {
   final int number;
 
   const AddEvent(this.number);
+}
+
+class IncrementEventHandler extends EventHandler<int, IncrementEvent> {
+  @override
+  FutureOr<void> call(IncrementEvent event) {
+    Future.delayed(Duration(seconds: 1));
+    setState(getState() + 1);
+  }
+}
+
+class DecrementEventHandler extends EventHandler<int, DecrementEvent> {
+  @override
+  bool callWhen(int state) => state > 0; // can not decrement below 0
+
+  @override
+  FutureOr<void> call(DecrementEvent event) {
+    setState(getState() - 1);
+  }
+}
+
+class AddHandler extends EventHandler<int, AddEvent> {
+  @override
+  Future<void> call(AddEvent event) async {
+    for (var i = 0; i < event.number; i++) {
+      await Future.delayed(Duration(milliseconds: 100));
+      final newState = getState() + 1;
+      setState(newState);
+    }
+  }
 }