+++ examples/todo_flutter/lib/helpers/object_ext.dart
@@ -3,4 +3,6 @@ extension ObjectExt on Object {
     if (this is T) return this as T;
     return null;
   }
+
+  T cast<T>() => this as T;
 }